Grundgerüst für Zweites Fenster erstellt und Loading Screen erstellt
