+++ MaxTemp/LoadingScreen.xaml
@@ -1,0 +1,12 @@
+﻿<Window x:Class="MaxTemp.LoadingScreen"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:MaxTemp"
+        mc:Ignorable="d"
+        Title="LoadingScreen" Height="450" Width="800">
+    <Grid>
+        
+    </Grid>
+</Window>
