+++ MaxTemp/ResultWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="MaxTemp.MainWindow"
+﻿<Window x:Class="MaxTemp.ResultWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -21,8 +21,7 @@
                     <RowDefinition Height="1*" />
                 </Grid.RowDefinitions>
 
-                <Label Grid.Row="0" Name="lblAufforderung" Margin="10" FontSize="16" Foreground="Black" HorizontalContentAlignment="Center">Bitte klicken, um Temperaturen auszuwerten.</Label>
-                <Button Grid.Row="0" Name="btnAuswerten" Click="BtnAuswerten_Click" Margin="10,40,10,10" FontSize="14" Background="Silver" Foreground="Black">Auswerten</Button>
+                <Label Grid.Row="0" Name="lblAufforderung" Margin="10" FontSize="16" Foreground="Black" HorizontalContentAlignment="Center">Hier ist die Auswertung.</Label>
 
                 <StackPanel Grid.Row="1" Margin="10">
                     <Label x:Name="lblAusgabe" Content="Höchste Temperatur:" FontSize="14" Foreground="Black"/>
